+++ .github/workflows/main.yml
@@ -28,7 +28,7 @@ jobs:
       - name: Build with Gradle
         run: ./gradlew clean build
       - name: Execute jacocoTestReport
-        run: ./gradlew test generateMergedReport
+        run: ./gradlew test generateMergedReport --stacktrace
       - name: copy the reportTest to codeCoverage
         run: cp build/reports/jacoco/generateMergedReport/generateMergedReport.xml jacoco.xml || echo "Code coverage failed"
       - name: Run sonarqube
