+++ examples/mq-mvc-app/src/main/java/co/com/bancolombia/jms/sample/noreactive/app/config/UseCasesConfig.java
@@ -1,11 +1,11 @@
-package co.com.bancolombia.jms.sample.app.config;
+package co.com.bancolombia.jms.sample.noreactive.app.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.FilterType;
 
 @Configuration
-@ComponentScan(basePackages = "co.com.bancolombia.jms.sample.domain.usecase",
+@ComponentScan(basePackages = "co.com.bancolombia.jms.sample.noreactive.domain.usecase",
         includeFilters = {
                 @ComponentScan.Filter(type = FilterType.REGEX, pattern = "^.+UseCase$")
         },
