+++ examples/mq-mvc-app/src/main/java/co/com/bancolombia/jms/sample/noreactive/domain/exceptions/RelatedMessageNotFoundException.java
@@ -1,4 +1,4 @@
-package co.com.bancolombia.jms.sample.domain.exceptions;
+package co.com.bancolombia.jms.sample.noreactive.domain.exceptions;
 
 public class RelatedMessageNotFoundException extends RuntimeException {
 }
