+++ examples/mq-mvc-app/src/main/java/co/com/bancolombia/jms/sample/noreactive/domain/model/RequestGateway.java
@@ -1,4 +1,4 @@
-package co.com.bancolombia.jms.sample.domain.model;
+package co.com.bancolombia.jms.sample.noreactive.domain.model;
 
 public interface RequestGateway {
     String send(Request request);
