+++ examples/mq-mvc-app/src/main/java/co/com/bancolombia/jms/sample/noreactive/entrypoints/MyMQListener.java
@@ -1,9 +1,9 @@
-package co.com.bancolombia.jms.sample.entrypoints;
+package co.com.bancolombia.jms.sample.noreactive.entrypoints;
 
 import co.com.bancolombia.commons.jms.mq.MQListener;
-import co.com.bancolombia.jms.sample.domain.model.Request;
-import co.com.bancolombia.jms.sample.domain.model.Result;
-import co.com.bancolombia.jms.sample.domain.usecase.ReplyRouterUseCase;
+import co.com.bancolombia.jms.sample.noreactive.domain.model.Request;
+import co.com.bancolombia.jms.sample.noreactive.domain.model.Result;
+import co.com.bancolombia.jms.sample.noreactive.domain.usecase.ReplyRouterUseCase;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.AllArgsConstructor;
