+++ examples/mq-mvc-app/src/main/java/co/com/bancolombia/jms/sample/noreactive/entrypoints/Rest.java
@@ -1,7 +1,7 @@
-package co.com.bancolombia.jms.sample.entrypoints;
+package co.com.bancolombia.jms.sample.noreactive.entrypoints;
 
-import co.com.bancolombia.jms.sample.domain.model.Result;
-import co.com.bancolombia.jms.sample.domain.usecase.SampleUseCase;
+import co.com.bancolombia.jms.sample.noreactive.domain.model.Result;
+import co.com.bancolombia.jms.sample.noreactive.domain.usecase.SampleUseCase;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
