+++ gradle.properties
@@ -1,4 +1,4 @@
-version=0.0.1
+version=0.0.2
 springBootVersion=2.4.8
 gradleVersionsVersion=0.28.0
 mqJMSVersion=2.5.0
